pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "football-tracking-demo"
version = "0.1.0"
description = "Football player detection and tracking demo using YOLOv8 and ByteTrack"
authors = [{ name = "Marco Willi", email = "marco.willi@fhnw.ch" }]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    # Core dependencies
    "numpy",
    "opencv-python",
    "PyYAML",

    # Computer Vision and ML
    "ultralytics", # YOLOv8 with built-in PyTorch
    "supervision", # ByteTrack tracker

    # Visualization
    "matplotlib",
    "Pillow",

    # Video processing
    "imageio",
    "imageio-ffmpeg",

    # Utilities
    "tqdm",
    "yt-dlp", # For video download
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff>=0.15.1",
    "pre-commit",
    "ipykernel",    # For Jupyter notebooks
    "jupyterlab",
    "pyrootutils",
]

[tool.ruff]
exclude = [".vscode", ".devcontainer"]

[tool.ruff.lint.per-file-ignores]
"notebooks/*" = ["E402"]
